finetune/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "one-ai-finetune"
version = "0.1.0"
description = "LoRA/QLoRA fine-tuning pipeline and evaluation for OpenNebula AI Assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openai>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
train = [
    "torch>=2.1",
    "transformers>=4.36",
    "peft>=0.7",
    "trl>=0.7",
    "bitsandbytes>=0.41",
    "datasets>=2.14",
    "accelerate>=0.25",
]
wandb = ["wandb>=0.16"]
dev = [
    "pytest>=7.0",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/one_ai_finetune"]

[tool.ruff]
line-length = 100
target-version = "py310"
